# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tinder-like Dating App API",
        "description": "RESTful API for a Tinder-like dating application with user profiles, recommendations, and matching functionality",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/people/recommended": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Get recommended users",
                "description": "Get recommended people for the authenticated user.",
                "operationId": "727fc82d8475a5aa8110ce939f1aa915",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Filter by gender",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "male",
                                "female",
                                "other"
                            ]
                        }
                    },
                    {
                        "name": "min_age",
                        "in": "query",
                        "description": "Minimum age",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 18
                        }
                    },
                    {
                        "name": "max_age",
                        "in": "query",
                        "description": "Maximum age",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recommended users retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "meta": {
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 20
                                                        },
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 150
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 8
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/people/{id}/like": {
            "post": {
                "tags": [
                    "People"
                ],
                "summary": "Like a user",
                "description": "Like a user.",
                "operationId": "f948dd4d5c3dd79ae7ad0e405b12a500",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID to like",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 123
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully liked user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully liked user"
                                        },
                                        "data": {
                                            "properties": {
                                                "liked_user_id": {
                                                    "type": "integer",
                                                    "example": 123
                                                },
                                                "liked_at": {
                                                    "type": "string",
                                                    "example": "2025-11-04T12:45:30Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cannot like yourself"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "409": {
                        "description": "Already liked this user"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/people/{id}/dislike": {
            "post": {
                "tags": [
                    "People"
                ],
                "summary": "Dislike a user",
                "description": "Dislike a user.",
                "operationId": "0c1b052305229a3750fded98090476c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID to dislike",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 123
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully disliked user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully disliked user"
                                        },
                                        "data": {
                                            "properties": {
                                                "disliked_user_id": {
                                                    "type": "integer",
                                                    "example": 123
                                                },
                                                "disliked_at": {
                                                    "type": "string",
                                                    "example": "2025-11-04T12:46:15Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cannot dislike yourself"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "409": {
                        "description": "Already disliked this user"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/people/liked": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Get liked users",
                "description": "Get users liked by the authenticated user.",
                "operationId": "a16694abd3d80a9ed4a65dc8a1eb110d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liked users retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "meta": {
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 20
                                                        },
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 45
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 3
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}